front/EXAMPLE_cloudacademy_next/fastapi-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt[crypto]==2.8.0
requests==2.31.0
boto3==1.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0